[wizard] Move get_username to helpers module
